// src/app/main/apps/todo/models/task.ts
export class Task {
    id: number;
    taskName: string;
    startDate: string;
    endDate: string;
    progress: number;
    duration: number;
    completedDate: string;
    createdDate: string;
    createdBy: string;
    projectName: string;
    taskResourceId: number;
    isInProgress: boolean;
    isCompleted: boolean;
    allocatedHours: number;
}